fix(50-06): remove redundant uint32_t cast on s_sprite_material.id
The .id field of nt_material_t is already uint32_t (engine/material/
nt_material.h:19). Cast was triggering both readability-redundant-casting
and google-readability-casting clang-tidy errors.

Surfaced post-Wave-3 by IDE LSP. Direct clang-tidy now clean. Build +
ctest 33/33 green.

Note: remaining IDE diagnostics in test_bunnymark_physics.c (file not
found, undeclared identifiers) and unused-includes warnings in main.c
are clangd LSP-only — direct clang-tidy is clean and the test target
builds + runs. Will resolve on next clangd reload.

+++ examples/bunnymark/main.c
@@ -309,7 +309,7 @@ static void frame(void) {
             uint32_t page_tex_id = nt_resource_get(page_res);
             s_items[i].sort_key = 0; /* unsorted in Bunnymark; renderer ignores */
             s_items[i].entity = s_entities[i].id;
-            s_items[i].batch_key = (page_tex_id << 16) | (uint32_t)s_sprite_material.id;
+            s_items[i].batch_key = (page_tex_id << 16) | s_sprite_material.id;
         }
         nt_sprite_renderer_draw_list(s_items, s_bunny_count);
         // #endregion
